--- a/python/ray/data/context.py
+++ b/python/ray/data/context.py
@@ -28,6 +28,9 @@ DEFAULT_OPTIMIZE_FUSE_READ_STAGES = True
 # Whether to furthermore fuse prior map tasks with shuffle stages.
 DEFAULT_OPTIMIZE_FUSE_SHUFFLE_STAGES = True
 
+# Wether to use actor based block prefetcher.
+DEFAULT_ACTOR_PREFETCHER_ENABLED = True
+
 
 @DeveloperAPI
 class DatasetContext:
@@ -46,6 +49,7 @@ class DatasetContext:
         optimize_fuse_stages: bool,
         optimize_fuse_read_stages: bool,
         optimize_fuse_shuffle_stages: bool,
+        actor_prefetcher_enabled: bool,
     ):
         """Private constructor (use get_current() instead)."""
         self.block_owner = block_owner
@@ -55,6 +59,7 @@ class DatasetContext:
         self.optimize_fuse_stages = optimize_fuse_stages
         self.optimize_fuse_read_stages = optimize_fuse_read_stages
         self.optimize_fuse_shuffle_stages = optimize_fuse_shuffle_stages
+        self.actor_prefetcher_enabled = actor_prefetcher_enabled
 
     @staticmethod
     def get_current() -> "DatasetContext":
@@ -76,6 +81,7 @@ class DatasetContext:
                     optimize_fuse_stages=DEFAULT_OPTIMIZE_FUSE_STAGES,
                     optimize_fuse_read_stages=DEFAULT_OPTIMIZE_FUSE_READ_STAGES,
                     optimize_fuse_shuffle_stages=DEFAULT_OPTIMIZE_FUSE_SHUFFLE_STAGES,
+                    actor_prefetcher_enabled=DEFAULT_ACTOR_PREFETCHER_ENABLED,
                 )
 
             if (