--- a/python/ray/data/impl/block_batching.py
+++ b/python/ray/data/impl/block_batching.py
@@ -64,7 +64,8 @@ def batch_blocks(
     for block_window in _sliding_window(blocks, prefetch_blocks + 1):
         block_window = list(block_window)
         with stats.iter_wait_s.timer():
-            ray.wait(block_window, num_returns=1, fetch_local=True)
+            prefercher = get_or_create_prefetcher()
+            prefercher.prefetch.remote(*block_window)
         yield from batch_block(block_window[0])
 
     # Consume remainder of final block window.
@@ -128,3 +129,26 @@ def _sliding_window(iterable: Iterable, n: int):
     for elem in it:
         window.append(elem)
         yield tuple(window)
+
+
+@ray.remote(num_cpus=0, placement_group=None)
+class _BlockPretcher:
+    def prefetch(self, *args):
+        pass
+
+def get_or_create_prefetcher():
+    ip_address = ray.util.get_node_ip_address()
+    actor_name = f"dataset-prefetcher-{ip_address}"
+    try:
+        prefetcher = ray.get_actor(
+            actor_name,
+            namespace="dataset",
+        )
+    except ValueError:
+        prefetcher = _BlockPretcher.options(
+            resources={"node:{}".format(ip_address): 0.0001},
+            name=actor_name,
+            namespace="dataset",
+            lifetime="detached",
+        ).remote()
+    return prefetcher