--- a/python/ray/data/impl/block_batching.py
+++ b/python/ray/data/impl/block_batching.py
@@ -69,6 +69,9 @@ def batch_blocks(
         block_window = list(block_window)
         with stats.iter_wait_s.timer():
             context = DatasetContext.get_current()
+            # ray.wait doesn't work as expected so that we use
+            # the actor based prefetcher as a work around. Read
+            # https://github.com/ray-project/ray/issues/23983 for details.
             if len(block_window) > 1 and context.actor_prefetcher_enabled:
                 prefetcher = get_or_create_prefetcher()
                 prefetcher.prefetch.remote(*block_window)