<!-- Thank you for your contribution! Please review https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR. If you don't have the access to it, we will shortly find a reviewer and assign them to your PR. -->

## Why are these changes needed?

The prefetch_blocks implementation doesn't work as expected. Due to `ray.wait()` doesn't given us fine grained control, today we block waiting any of the block returns. As I read the code, it may or may not actually fetching all the blocks.

A better way to ensure prefetching not blocking is to use ray remote function call, which is not blocking and ensures the blocks are fetched eventually.

The evaluation shows the actor based prefetcher could achieve **perfect prefetching** if the prefetch speed is faster than consuming speed.

## Evaluation
Preliminary test shows the actor based prefetcher improve the throughtput of [following script](https://gist.github.com/scv119/d62d141fe00a50c503ff2530731a81eb) on a [5 cpu node 1 gpu node cluster](https://gist.github.com/scv119/424d8bb6901e65d8acf173eb72bc1254) , which simulate ingesting validation dataset by
```
1. load 80GB data on 5 cpu nodes.
2. ingest the data on gpu nodes with ds.iter_batch(prefetching_blocks=1, batch_size=250000), where fo reach batch we sleep for a small time period to simulate validation.
3. we measure and report the total time of step 2 (validation)
```

### 0.01 second per batch validation time:

If we simulate per-batch validation time to 0.01 seconds:

 validation time: **88.94s -> 68.13s**

ray.wait prefetching:
```
(consume pid=11517, ip=10.0.1.12) == Pipeline Window 0 ==
(consume pid=11517, ip=10.0.1.12) Stage 1 read: 48/48 blocks executed in 26.37s
(consume pid=11517, ip=10.0.1.12) * Remote wall time: 5.12s min, 24.05s max, 13.77s mean, 660.84s total
(consume pid=11517, ip=10.0.1.12) * Remote cpu time: 4.35s min, 5.33s max, 4.92s mean, 235.99s total
(consume pid=11517, ip=10.0.1.12) * Output num rows: 10000000 min, 10000000 max, 10000000 mean, 480000000 total
(consume pid=11517, ip=10.0.1.12) * Output size bytes: 1787500000 min, 1787500000 max, 1787500000 mean, 85800000000 total
(consume pid=11517, ip=10.0.1.12) * Tasks per node: 9 min, 10 max, 9 mean; 5 nodes used
(consume pid=11517, ip=10.0.1.12)
(consume pid=11517, ip=10.0.1.12) Stage 2 map_batches: 48/48 blocks executed in 60.55s
(consume pid=11517, ip=10.0.1.12) * Remote wall time: 11.41s min, 16.25s max, 14.01s mean, 672.57s total
(consume pid=11517, ip=10.0.1.12) * Remote cpu time: 11.95s min, 16.8s max, 14.58s mean, 699.93s total
(consume pid=11517, ip=10.0.1.12) * Output num rows: 10000000 min, 10000000 max, 10000000 mean, 480000000 total
(consume pid=11517, ip=10.0.1.12) * Output size bytes: 1680000128 min, 1680000128 max, 1680000128 mean, 80640006144 total
(consume pid=11517, ip=10.0.1.12) * Tasks per node: 9 min, 10 max, 9 mean; 5 nodes used
(consume pid=11517, ip=10.0.1.12)
(consume pid=11517, ip=10.0.1.12) Dataset iterator time breakdown:
(consume pid=11517, ip=10.0.1.12) * In ray.wait(): 41.07s
(consume pid=11517, ip=10.0.1.12) * In ray.get(): 24.78s
(consume pid=11517, ip=10.0.1.12) * In format_batch(): 2.9s
(consume pid=11517, ip=10.0.1.12) * In user code: 19.4s
(consume pid=11517, ip=10.0.1.12) * Total time: 88.22s
(consume pid=11517, ip=10.0.1.12)
(consume pid=11517, ip=10.0.1.12) ##### Overall Pipeline Time Breakdown #####
(consume pid=11517, ip=10.0.1.12) * Time in dataset iterator: 68.82s
(consume pid=11517, ip=10.0.1.12) * Time in user code: 19.39s
(consume pid=11517, ip=10.0.1.12) * Total time: 88.94s
(consume pid=11517, ip=10.0.1.12)
```
actor based prefetching

```
(consume pid=11209, ip=10.0.1.12) == Pipeline Window 0 ==
(consume pid=11209, ip=10.0.1.12) Stage 1 read: 48/48 blocks executed in 32.15s
(consume pid=11209, ip=10.0.1.12) * Remote wall time: 5.88s min, 30.26s max, 15.79s mean, 758.15s total
(consume pid=11209, ip=10.0.1.12) * Remote cpu time: 4.54s min, 5.84s max, 5.01s mean, 240.63s total
(consume pid=11209, ip=10.0.1.12) * Output num rows: 10000000 min, 10000000 max, 10000000 mean, 480000000 total
(consume pid=11209, ip=10.0.1.12) * Output size bytes: 1787500000 min, 1787500000 max, 1787500000 mean, 85800000000 total
(consume pid=11209, ip=10.0.1.12) * Tasks per node: 9 min, 10 max, 9 mean; 5 nodes used
(consume pid=11209, ip=10.0.1.12)
(consume pid=11209, ip=10.0.1.12) Stage 2 map_batches: 48/48 blocks executed in 63.56s
(consume pid=11209, ip=10.0.1.12) * Remote wall time: 11.48s min, 17.06s max, 14.36s mean, 689.4s total
(consume pid=11209, ip=10.0.1.12) * Remote cpu time: 12.08s min, 17.57s max, 14.94s mean, 717.03s total
(consume pid=11209, ip=10.0.1.12) * Output num rows: 10000000 min, 10000000 max, 10000000 mean, 480000000 total
(consume pid=11209, ip=10.0.1.12) * Output size bytes: 1680000128 min, 1680000128 max, 1680000128 mean, 80640006144 total
(consume pid=11209, ip=10.0.1.12) * Tasks per node: 9 min, 10 max, 9 mean; 5 nodes used
(consume pid=11209, ip=10.0.1.12)
(consume pid=11209, ip=10.0.1.12) Dataset iterator time breakdown:
(consume pid=11209, ip=10.0.1.12) * In ray.wait(): 33.04ms
(consume pid=11209, ip=10.0.1.12) * In ray.get(): 43.91s
(consume pid=11209, ip=10.0.1.12) * In format_batch(): 3.85s
(consume pid=11209, ip=10.0.1.12) * In user code: 19.52s
(consume pid=11209, ip=10.0.1.12) * Total time: 67.41s
(consume pid=11209, ip=10.0.1.12)
(consume pid=11209, ip=10.0.1.12) ##### Overall Pipeline Time Breakdown #####
(consume pid=11209, ip=10.0.1.12) * Time in dataset iterator: 47.9s
(consume pid=11209, ip=10.0.1.12) * Time in user code: 19.49s
(consume pid=11209, ip=10.0.1.12) * Total time: 68.13s
```

### 0.05 second per batch validation time:

If we simulate per-batch validation time to 0.05 seconds:

**validation time: 165.99 -> 104.1s**

Actually as we can see from the experiment, we achieved **perfect prefetching as the ray.get/ray.wait time is close to 0.**

ray.wait prefetching:
```
(consume pid=11990, ip=10.0.1.12) == Pipeline Window 0 ==
(consume pid=11990, ip=10.0.1.12) Stage 1 read: 48/48 blocks executed in 30.24s
(consume pid=11990, ip=10.0.1.12) * Remote wall time: 7.08s min, 28.37s max, 13.67s mean, 656.38s total
(consume pid=11990, ip=10.0.1.12) * Remote cpu time: 4.37s min, 5.4s max, 4.88s mean, 234.34s total
(consume pid=11990, ip=10.0.1.12) * Output num rows: 10000000 min, 10000000 max, 10000000 mean, 480000000 total
(consume pid=11990, ip=10.0.1.12) * Output size bytes: 1787500000 min, 1787500000 max, 1787500000 mean, 85800000000 total
(consume pid=11990, ip=10.0.1.12) * Tasks per node: 9 min, 10 max, 9 mean; 5 nodes used
(consume pid=11990, ip=10.0.1.12)
(consume pid=11990, ip=10.0.1.12) Stage 2 map_batches: 48/48 blocks executed in 58.15s
(consume pid=11990, ip=10.0.1.12) * Remote wall time: 11.5s min, 16.48s max, 13.9s mean, 667.43s total
(consume pid=11990, ip=10.0.1.12) * Remote cpu time: 12.09s min, 17.02s max, 14.48s mean, 694.83s total
(consume pid=11990, ip=10.0.1.12) * Output num rows: 10000000 min, 10000000 max, 10000000 mean, 480000000 total
(consume pid=11990, ip=10.0.1.12) * Output size bytes: 1680000128 min, 1680000128 max, 1680000128 mean, 80640006144 total
(consume pid=11990, ip=10.0.1.12) * Tasks per node: 9 min, 10 max, 9 mean; 5 nodes used
(consume pid=11990, ip=10.0.1.12)
(consume pid=11990, ip=10.0.1.12) Dataset iterator time breakdown:
(consume pid=11990, ip=10.0.1.12) * In ray.wait(): 43.73s
(consume pid=11990, ip=10.0.1.12) * In ray.get(): 22.03s
(consume pid=11990, ip=10.0.1.12) * In format_batch(): 3.14s
(consume pid=11990, ip=10.0.1.12) * In user code: 96.28s
(consume pid=11990, ip=10.0.1.12) * Total time: 165.27s
(consume pid=11990, ip=10.0.1.12)
(consume pid=11990, ip=10.0.1.12) ##### Overall Pipeline Time Breakdown #####
(consume pid=11990, ip=10.0.1.12) * Time in dataset iterator: 69.0s
(consume pid=11990, ip=10.0.1.12) * Time in user code: 96.26s
(consume pid=11990, ip=10.0.1.12) * Total time: 165.99
```
actor based prefetching
```
(consume pid=11780, ip=10.0.1.12) == Pipeline Window 0 ==
(consume pid=11780, ip=10.0.1.12) Stage 1 read: 48/48 blocks executed in 30.43s
(consume pid=11780, ip=10.0.1.12) * Remote wall time: 5.98s min, 28.4s max, 14.35s mean, 688.75s total
(consume pid=11780, ip=10.0.1.12) * Remote cpu time: 4.53s min, 5.7s max, 4.91s mean, 235.8s total
(consume pid=11780, ip=10.0.1.12) * Output num rows: 10000000 min, 10000000 max, 10000000 mean, 480000000 total
(consume pid=11780, ip=10.0.1.12) * Output size bytes: 1787500000 min, 1787500000 max, 1787500000 mean, 85800000000 total
(consume pid=11780, ip=10.0.1.12) * Tasks per node: 9 min, 10 max, 9 mean; 5 nodes used
(consume pid=11780, ip=10.0.1.12)
(consume pid=11780, ip=10.0.1.12) Stage 2 map_batches: 48/48 blocks executed in 61.91s
(consume pid=11780, ip=10.0.1.12) * Remote wall time: 11.27s min, 15.14s max, 13.62s mean, 653.93s total
(consume pid=11780, ip=10.0.1.12) * Remote cpu time: 11.89s min, 15.93s max, 14.22s mean, 682.59s total
(consume pid=11780, ip=10.0.1.12) * Output num rows: 10000000 min, 10000000 max, 10000000 mean, 480000000 total
(consume pid=11780, ip=10.0.1.12) * Output size bytes: 1680000128 min, 1680000128 max, 1680000128 mean, 80640006144 total
(consume pid=11780, ip=10.0.1.12) * Tasks per node: 9 min, 10 max, 9 mean; 5 nodes used
(consume pid=11780, ip=10.0.1.12)
(consume pid=11780, ip=10.0.1.12) Dataset iterator time breakdown:
(consume pid=11780, ip=10.0.1.12) * In ray.wait(): 24.93ms
(consume pid=11780, ip=10.0.1.12) * In ray.get(): 2.93s
(consume pid=11780, ip=10.0.1.12) * In format_batch(): 4.0s
(consume pid=11780, ip=10.0.1.12) * In user code: 96.3s
(consume pid=11780, ip=10.0.1.12) * Total time: 103.38s
(consume pid=11780, ip=10.0.1.12)
(consume pid=11780, ip=10.0.1.12) ##### Overall Pipeline Time Breakdown #####
(consume pid=11780, ip=10.0.1.12) * Time in dataset iterator: 7.09s
(consume pid=11780, ip=10.0.1.12) * Time in user code: 96.26s
(consume pid=11780, ip=10.0.1.12) * Total time: 104.1s
```

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for https://docs.ray.io/en/master/.
- [ ] I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [x] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(