@clarkzinzow 2022-04-16
Awesome, great benchmarking!

It seems like the underlying problem with the current implementation is that `ray.wait()` isn't working as expected. For `ray.wait(refs, num_returns=1, fetch_local=True)`, a pull request should be issued for every object in `ref`, and the `ray.wait()` call shouldn't return until one of the objects is local. The actor-based prefetcher is doing the same thing, triggering a pull request for every object in `refs`, but isn't blocking until one of them is local. Ultimately, I'd expect the time in prefetching + the time in `ray.get()` to be ~ the same across these two approaches, with the `ray.wait()` approach being slightly faster due to the lack of task submission overhead. I think that the only difference in the created pull requests are the rough priority of `ray.wait()` calls vs task arguments under memory pressure, which I assume isn't relevant in this case.

I'm assuming that we don't know why `ray.wait()` is not behaving as expected here?