@clarkzinzow 2022-04-16
Also, if we were to bypass `ray.wait()` in our prefetching, instead of creating a detached actor for each consumer node, I'd vote for issuing `ray.get` prefetches in a background thread within the consumer process (created in `.iter_batches()`), where the main thread pulls these prefetch blocks via a threading queue. Ludwig current does this themselves since they ran into bugs with our old prefetching implementation, and they've had great results.

We were already planning to do this in order to be able to pipeline some last-mile per-batch transformations (Torch tensor conversion, local shuffling, device transfer, etc.) with training, so this implementation would stick even if we fixed `ray.wait`.