@scv119 2022-04-18
> I'm assuming that we don't know why ray.wait() is not behaving as expected here?

i think ray.wait behavior is not well defined and can only be reasoned by reading the code. 

> Also, if we were to bypass ray.wait() in our prefetching, instead of creating a detached actor for each consumer node, I'd vote for issuing ray.get prefetches in a background thread within the consumer process (created in .iter_batches()),

Do you have code snippet on how to do that? Is calling `ray.get` in a separate thread defined behavior?