@stephanie-wang 2022-04-18
Agreed, this sounds like a bug in `ray.wait(fetch_local=True)`. Could you open an issue with the findings here? As @scv119 says, we should also better define the semantics for this call. It should act the same way as the `ray.get` hack here as long as there is memory availability, but yes, this is not documented right now.