@clarkzinzow 2022-04-24
@scv119 I'll make a pass at a background thread-based solution on Monday, and if that doesn't pan out, we can merge this actor-based solution.