@jianoaix 2022-04-25
This PR looks good to me. Actually, from my understanding, using Actors is cleaner since it's a higher level abstraction than using raw threads. If the performance diff is small, Actor should be preferred. Anything I am missing?