@clarkzinzow 2022-04-25
@jianoaix There are multiple issues with this kind of actor-based prefetching, detailed [here](https://github.com/ray-project/ray/issues/24158).